+++ src/Products/PloneMeeting/migrations/migrate_to_4215.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from imio.helpers.setup import load_type_from_package
+from imio.webspellchecker.config import get_disable_autosearch_in
+from imio.webspellchecker.config import set_disable_autosearch_in
 from Products.PloneMeeting.migrations import logger
 from Products.PloneMeeting.migrations import Migrator
 
@@ -27,6 +29,18 @@ class Migrate_To_4215(Migrator):
                 cfg.registerPortalTypes()
         logger.info('Done.')
 
+    def _fixWSCConfigAndCleanBrokenAnnexes(self):
+        """WSC could sometimes break quickupload leading to broken annexes.
+           Adapt WSC config so it is no more used in quickupload and clean broken annexes."""
+        logger.info("Updating WSC config and removing broken annexes...")
+        if self.portal.portal_quickinstaller.isProductInstalled('imio.webspellchecker'):
+            # disable WSC in input and textarea
+            if not get_disable_autosearch_in():
+                set_disable_autosearch_in(u'["input", "textarea"]')
+            # remove broken annexes
+            self._removeBrokenAnnexes()
+        logger.info('Done.')
+
     def run(self, extra_omitted=[], from_migration_to_4200=False):
 
         logger.info('Migrating to PloneMeeting 4215...')
@@ -41,6 +55,7 @@ class Migrate_To_4215(Migrator):
             self.ps.runImportStepFromProfile('profile-Products.PloneMeeting:default', 'viewlets')
         self._updateConfigCustomAdvisersDataGrid()
         self._reloadMeetingConfigsForItemWorkflows()
+        self._fixWSCConfigAndCleanBrokenAnnexes()
         logger.info('Migrating to PloneMeeting 4215... Done.')
 
 
@@ -49,7 +64,8 @@ def migrate(context):
 
        1) Upgrade all and make sure documentgenerator overrides are re-applied;
        2) Update MeetingConfig.customAdvisers to add new column "is_delay_calendar_days";
-       3) Reload MeetingConfigs if using "return_to_proposing_group" with validation.
+       3) Reload MeetingConfigs if using "return_to_proposing_group" with validation;
+       4) Fix WSC config and remove broken annexes.
     '''
     migrator = Migrate_To_4215(context)
     migrator.run()